Move main to separate file
Move logic related to Main which is not covered by test to a separate file and exclude it from coverage
