+++ .codecov.yml
@@ -1,0 +1,3 @@
+
+ignore:
+  - "**/*Main.scala"
